Add basic actions (#1)
This is the first commit, which adds some basic actions for the
recloneui. It is possible to select a configured remote environment and
to navigate through the files and folders of the selected remote.

A selected file can be copied to the current local directory or it can
be deleted.