Use tview instead of prompt
Instead of the prompt package we are now using tview. This allows us to
render the files and folders in a table. We can now also switch between
two tables, so that a user can copy a file/folder in one view and paste
it in the other one.

To copy, paste or delete a file/folder the user can now use the
following keys instead of enter: c (copy), p (paste) and dd (delete).