Add support for filtering
This commit adds support for filtering via the "--min-age", "--max-age",
"--min-size" and "--max-size" flags. These values can only be set via
flags and they are applied for all remotes. The default value for all
flags is "off" and they are using the same format as the corresponding
rclone flags.

NOTE: Maybe we can add a modal to set these flags during the runtime of
the rcloneui.